% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify_across.R
\name{dummify_across}
\alias{dummify_across}
\title{Make Multiple Dummy (Binary) Variables At Once}
\usage{
dummify_across(data, vars, reference = NULL)
}
\arguments{
\item{data}{A dataframe containing the variables to be dummied.}

\item{vars}{A list of column names or indices. Accepts variables integer, string, and factor values.}

\item{reference}{If TRUE, it will leave the first value observed across each variable out as a
reference category. Defaults to NULL, which leaves no variables as references.}
}
\value{
The inputted dataframe with additional columns containing all of the dummified variables
}
\description{
This function allows you to make dummy variables across many variables at once. For example,
if you have two variables called sex (with values "male","female","other") and marital status
(with values "single", "married", "divorced", "widowed"), this function creates dummy
variables for both sex and marital status based off of their unique values. This function is
a wrapper of \code{dummify()} that allows multiple variables to be passed in the same function
call. Currently, automatic renaming of variables and manual selection of reference categories
is limited to \code{dummify()}.
}
\examples{
install.packages("palmerpenguins")

penguins<-palmerpenguins::penguins

dummify_across(penguins, c("species","island","sex"))
dummify_across(penguins, c(1,2,7))
dummify_across(penguins, c("species","island","sex"), reference = TRUE)

}
\seealso{
\code{\link{dummify}}
}
\keyword{dummy}
